new-utility-agent/requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
langgraph
google-generativeai
langchain-google-genai

# Beckn/UEI specific, if needed later
# py-beckn-client 